Separate global Discord logic from agile implementation
When we want to introduce new ways of interacting with Discord, we need to separate the current 'agile' implementation with everything else.
This folder structure makes much more sense therefore.
