+++ api/src/discord/agile/commands/solveTask.ts
@@ -4,11 +4,11 @@ import {
   Client,
   CommandInteraction,
 } from "discord.js";
-import { Command } from "./command";
-import { setFlagForChallengeId } from "../database/tasks";
-import { handleTaskSolved } from "../../plugins/discordHooks";
-import { getUserByDiscordId } from "../database/users";
-import { getCurrentTaskChannelFromDiscord } from "../utils/channels";
+import { Command } from "../../interfaces/command";
+import { setFlagForChallengeId } from "../../database/tasks";
+import { handleTaskSolved } from "../../../plugins/discordHooks";
+import { getUserByDiscordId } from "../../database/users";
+import { getCurrentTaskChannelFromDiscord } from "../../utils/channels";
 
 async function accessDenied(interaction: CommandInteraction) {
   await interaction.editReply({
