--- api/src/discord/interactions/interactions.ts
@@ -1,10 +1,0 @@
-import { DiscordButtonInteraction } from "./interaction";
-import { HandleArchiveCtfInteraction } from "../commands/archiveCtf";
-import { HandleCreateCtfInteraction } from "../commands/createCtf";
-import { HandleDeleteCtfInteraction } from "../commands/deleteCtf";
-
-export const Interactions: DiscordButtonInteraction[] = [
-  HandleCreateCtfInteraction,
-  HandleDeleteCtfInteraction,
-  HandleArchiveCtfInteraction,
-];
