+++ api/src/discord/interfaces/command.ts
