+++ api/src/discord/interfaces/interaction.ts
